requirements.txt
absl-py==2.2.0
accelerate==1.6.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.1
cupy-cuda12x==13.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
edge-tts==6.1.12
einops==0.8.0
environs==9.5.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.109.2
fastrlock==0.8.3
ffmpy==0.4.0
filelock==3.16.0
Flask==3.0.3
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
gradio==4.29.0
gradio_client==0.16.1
greenlet==3.1.0
grpcio==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.7
hyperframe==6.0.1
icecream==2.1.3
idna==3.9
importlib_metadata==8.5.0
importlib_resources==6.4.5
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.120
llama-index==0.10.10
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.1.2
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-milvus==0.1.3
llama-parse==0.4.9
llamaindex-py-client==0.1.19
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
minijinja==2.2.0
minio==7.2.8
modelscope==1.12.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.8.0
nebula3-python==3.5.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.45.0
orjson==3.10.7
oss2==2.19.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pansi==2020.7.3
pdfminer==20191125
pdfminer.six==20240706
pillow==10.4.0
platformdirs==4.3.3
prettytable==3.11.0
protobuf==5.28.1
psutil==6.0.0
putils==0.2.3
py2neo==2021.2.4
pyahocorasick==2.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pymilvus==2.3.6
pyparsing==3.1.4
pypdf==4.0.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.2
PyYAML==6.0.2
ragas==0.2.14
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
rich==13.8.1
rouge_score==0.1.2
rpds-py==0.20.0
ruff==0.6.5
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.1.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.36.3
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.2
tqdm==4.66.5
transformers==4.38.2
triton==2.2.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.27.1
watchdog==6.0.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.5.0
yapf==0.40.2
yarl==1.11.1
zhipuai==2.0.1.20240429
zipp==3.20.2
